requirements.txt
streamlit==1.15.2
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
prophet==1.0.1
holidays==0.11.3.1
tensorflow==2.6.0
Pillow==8.3.2
google-generativeai==0.2.0
python-dotenv==0.19.1
